Revert some not needed changes
