+++ include/osr/types.h
@@ -18,12 +18,6 @@
 
 namespace osr {
 
-struct routing_parameters {
-  float speed_{1.7F};
-};
-
-constexpr auto const kRoutingParameters = routing_parameters{};
-
 template <typename K, typename V, typename SizeType = cista::base_t<K>>
 using vecvec = cista::raw::vecvec<K, V, SizeType>;
